requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

requests==2.31.0
markdownify
azure-functions
azure-storage-blob
python-dotenv